LayoutTests/inspector/uncaught-dom3-exception-expected.txt

CONSOLE MESSAGE: line 9: HIERARCHY_REQUEST_ERR: DOM Exception 3: A Node was inserted somewhere it doesn't belong.
This tests that we are getting the correct message for DOM Exception 3: HIERARCHY_REQUEST_ERR.
